requirements.txt
absl-py==1.0.0
addict==2.4.0
affine==2.3.0
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
click-plugins==1.1.1
cligj==0.7.2
cycler==0.11.0
docker-pycreds==0.4.0
einops==0.4.0
entmax==1.0
Fiona==1.8.21
fonttools==4.29.1
frozenlist==1.3.0
fsspec==2022.1.0
future==0.18.2
geopandas==0.10.2
gitdb==4.0.9
GitPython==3.1.26
google-auth==2.6.0
google-auth-oauthlib==0.4.6
grpcio==1.43.0
idna==3.3
imageio==2.15.0
importlib-metadata==4.10.1
joblib==1.1.0
kiwisolver==1.3.2
kornia==0.6.3
Markdown==3.3.6
matplotlib==3.5.1
mlm-pytorch==0.1.0
multidict==6.0.2
munch==2.5.0
networkx==2.6.3
numpy==1.22.2
oauthlib==3.2.0
packaging==21.3
pandas==1.4.0
pathtools==0.1.2
Pillow==9.0.1
promise==2.3
protobuf==3.19.4
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.1
pyparsing==3.0.7
pyproj==3.3.0
python-dateutil==2.8.2
pytorch-lightning==1.5.10
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
rasterio==1.2.10
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.1
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
sentry-sdk==1.5.4
Shapely==1.8.0
shortuuid==1.0.8
six==1.16.0
sklearn==0.0
smmap==5.0.0
snuggs==1.4.7
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2022.2.9
torch==1.10.2
torchmetrics==0.7.1
tqdm==4.62.3
typing_extensions==4.0.1
urllib3==1.26.8
wandb==0.12.10
Werkzeug==2.0.3
x-transformers==0.23.0
yarl==1.7.2
yaspin==2.1.0
zipp==3.7.0
